Multiple tracking2
Loading the Video: The video is loaded, and two new variables, boundingBoxes and colors, are defined as empty lists.

Selecting Multiple Bounding Boxes: A while loop is used to select multiple bounding boxes in the first frame of the video. The cv2.selectROI function is used to select the region of interest.

Assigning Colors: Random colors are generated for each bounding box using randint(0, 255).

Interrupting the Loop: The loop can be interrupted by pressing the 'Q' or 'q' key, which breaks the loop.