MEANSHIFT Algorithm
What is Meanshift Algorithm?

The mean shift algorithm, which uses spatial analysis to find the maximum of a density function. The algorithm works by determining the weights of nearby points or pixels based on an average calculation using colors. It moves towards the area of maximum pixel density, where there are more pixels than in other parts of the image.

The use animations to demonstrate how the algorithm tracks a white pixel and moves towards it, showing the density function in action. The goal is to approximate the pixels according to the features of the initial object.

The concept of the Probability Density Function (PDF) is explained, which is used in the mean shift algorithm. A confidence map based on probabilities is created to approximate frames in a video, helping the algorithm identify the target object's location.

In summary, the object we want to track is in the central part where the density or number of pixels is greater.